Add LFO to FilterEffect. Closes #79

+++ libblankenhain/src/AbletonTrackMixerVolumeEffect.cpp
@@ -23,18 +23,18 @@ void AbletonTrackMixerVolumeEffect::process(Sample* buffer, size_t numberOfSampl
 		float fvol = volume.get();
 		if (fvol > 1.f)
 		{
-			float mult = -6.02413 + 2.71507e-4 * ((fvol - 1.f) * 22050.f);
+			float mult = -6.02413f + static_cast<float>(2.71507e-4) * ((fvol - 1.f) * 22050.f);
 			multiplier = Sample(aux::decibelToLinear(mult + 6.f));
 		}
 		else if (fvol >= 0.5f)
 		{
-			float mult2 = -16.05759 + 24.23926 * fvol - 8.20311 * fvol * (fvol);
+			float mult2 = -16.05759f + 24.23926f * fvol - 8.20311f * fvol * (fvol);
 			multiplier = Sample(aux::decibelToLinear(mult2));
 		}
 		else
 		{
 			float x = (fvol * 22050.f);
-			float mult = -67.17493 + 0.00855 * x - 4.70518e-7 * x * x + 1.01315e-11 * x * x * x;
+			float mult = -67.17493f + 0.00855f * x - static_cast<float>(4.70518e-7) * x * x + static_cast<float>(1.01315e-11) * x * x * x;
 			mult += 12.f;
 			if (mult > -6.f)
 				mult = -6.f;
